refactor: remove build tooling and use CDN modules
