+++ index.html
@@ -7,6 +7,12 @@
   </head>
   <body>
     <div id="root"></div>
-    <script type="module" src="/src/main.tsx"></script>
+    <script type="module" src="./main.js"></script>
+    <style>
+      .rec {
+        color: red;
+        margin-left: 8px;
+      }
+    </style>
   </body>
 </html>
