--- vite.config.ts
@@ -1,29 +1,0 @@
-import { defineConfig } from 'vite';
-import react from '@vitejs/plugin-react';
-import { VitePWA } from 'vite-plugin-pwa';
-
-export default defineConfig({
-  base: './',
-  plugins: [
-    react(),
-    VitePWA({
-      registerType: 'autoUpdate',
-      includeAssets: ['favicon.svg'],
-      manifest: {
-        name: 'Voice Pause Video',
-        short_name: 'VoicePause',
-        start_url: '.',
-        display: 'standalone',
-        background_color: '#ffffff',
-        description: 'Record voice overs while pausing video',
-        icons: [
-          {
-            src: 'favicon.svg',
-            sizes: 'any',
-            type: 'image/svg+xml'
-          }
-        ]
-      }
-    })
-  ]
-});
